Make server configuration script usable for VM and existing hosts.
It tells from the path of the script itself.

+++ scripts/configure-server.sh
@@ -1,25 +1,42 @@
 #!/bin/sh
 # Configure this (virtual) host as an eReuse PXE server.
 #
-# This is meant to be run when generating the server OVA.
+# This is meant to be run as ``/etc/rc.local`` when generating the server OVA,
+# then it takes over the whole host.
+#
+# You may also run it on an existing host, then it just adds the needed
+# configuration.
+
+if [ "$0" = /etc/rc.local ]; then
+    vm=yes  # take full control of virtual host
+else
+    vm=no  # just add eReuse configuration
+fi
 
 # Remove unnecessary packages.
-apt-get -qq purge @PKGS_TO_REMOVE@ thermald plymouth
+if [ $vm = yes ]; then
+    apt-get -qq purge @CASPER_PKGS@ thermald plymouth
+fi
 
 # Install needed packages.
-# Enable additional packages like VirtualBox's.
-sed -i -e 's/ main/ main multiverse/' /etc/apt/sources.list
+pkgs_to_install="
+    isc-dhcp-server tftpd-hpa pxelinux syslinux-common
+    nfs-kernel-server samba"
+if [ $vm = yes ]; then
+    # Enable VirtualBox's packages.
+    sed -i -e 's/ main/ main multiverse/' /etc/apt/sources.list
+    pkgs_to_install="$pkgs_to_install virtualbox-guest-dkms"
+fi
 apt-get -qq update
-apt-get -qq --no-install-recommends install \
-        virtualbox-guest-dkms \
-        isc-dhcp-server tftpd-hpa pxelinux syslinux-common \
-        nfs-kernel-server samba
+apt-get -qq --no-install-recommends install $pkgs_to_install
 
 # Enable VirtualBox's shared folder module.
-cat << 'EOF' > /etc/modules-load.d/ereuse.conf
+if [ $vm = yes ]; then
+    cat << 'EOF' > /etc/modules-load.d/ereuse.conf
 # To share eReuse's data folder with the VirtualBox host.
 vboxsf
 EOF
+fi
 
 # Configure the DHCP server.
 address=$(ip -4 addr show dev eth0 | sed -nr 's/.*\binet ([^/]+).*/\1/p')
@@ -44,6 +61,9 @@ subnet $netpfx24.0 netmask 255.255.255.0 {
   option broadcast-address $netpfx24.255;
 }
 EOF
+if [ $vm = no ]; then
+    service isc-dhcp-server restart
+fi
 
 # Configure the Samba server.
 cat << 'EOF' >> /etc/samba/smb.conf
@@ -57,6 +77,9 @@ cat << 'EOF' >> /etc/samba/smb.conf
    force user = ereuse
    force group = ereuse
 EOF
+if [ $vm = no ]; then
+    service samba restart
+fi
 
 # Configure PXE boot with TFTP.
 sed -i -e 's/\[::\]//' /etc/default/tftpd-hpa  # Ubuntu bug #1448500
@@ -71,20 +94,28 @@ timeout 50
 
 ###eReuse###
 EOF
-
-# Rebuild initramfs if missing.
-if [ ! -f /initrd.img ]; then
-    update-initramfs -u
+if [ $vm = no ]; then
+    service tftpd-hpa restart
 fi
 
-# Avoid attempting graphical login and get a console login prompt.
-sed -i -r -e 's/^(GRUB_HIDDEN_.*)/#\1/' -e 's/quiet splash/quiet/' /etc/default/grub
-# Fix boot loader.
-update-grub
-grub-install $(findmnt -no SOURCE / | sed -r 's/p?[0-9]+$//')
+# Configure boot.
+if [ $vm = yes ]; then
+    # Rebuild initramfs if missing.
+    if [ ! -f /initrd.img ]; then
+        update-initramfs -u
+    fi
+
+    # Avoid attempting graphical login and get a console login prompt.
+    sed -i -r -e 's/^(GRUB_HIDDEN_.*)/#\1/' -e 's/quiet splash/quiet/' /etc/default/grub
+    # Fix boot loader.
+    update-grub
+    grub-install $(findmnt -no SOURCE / | sed -r 's/p?[0-9]+$//')
+fi
 
 # Setup users.
-printf 'eReuse\neReuse\n' | passwd -q root
+if [ $vm = yes ]; then
+    printf 'eReuse\neReuse\n' | passwd -q root
+fi
 adduser --disabled-password --gecos eReuse ereuse
 printf 'ereuse\nereuse\n' | passwd -q ereuse
 
@@ -93,16 +124,23 @@ mkdir -m 0755 ~ereuse/data
 ln -s data/config.ini data/inventory ~ereuse  # compat links
 chown -R ereuse:ereuse ~ereuse
 mkdir -p /srv/ereuse-data
-cat << 'EOF' >> /etc/fstab
-# Mount the VirtualBox shared folder, then bind it to the usual place.
+if [ $vm = yes ]; then
+    cat << 'EOF' >> /etc/fstab
+# Mount the VirtualBox shared folder as the data directory.
 ereuse-data        /home/ereuse/data  vboxsf  rw,uid=ereuse,gid=ereuse,dmode=755,fmode=644  0  0
+EOF
+fi
+cat << 'EOF' >> /etc/fstab
+# Bind the data directory to the usual place.
 /home/ereuse/data  /srv/ereuse-data   none    bind  0  0
 EOF
+if [ $vm = no ]; then
+    mount -a
+fi
 
-# Cleanup.
-apt-get clean  # downloaded package files
-
-# Restore the original init script.
-mv /etc/rc.local.orig /etc/rc.local
-
-exec poweroff
+# Cleanup, restore the original init script and halt the machine.
+if [ $vm = yes ]; then
+    apt-get clean  # downloaded package files
+    mv /etc/rc.local.orig /etc/rc.local
+    exec poweroff
+fi
